small tweaks

+++ src/main/java/com/example/danhom1/UserStorage/UserStorageRepo.java
@@ -2,4 +2,4 @@ package com.example.danhom1.UserStorage;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserStorageRepo extends JpaRepository<UserStorage, Integer> { }
+public interface UserStorageRepo extends JpaRepository<UserStorage, Long> { }
